Move loading to own module
