+++ src/main.rs
@@ -1,5 +1,4 @@
 use anyhow::Result;
-use chrono::{NaiveDate, NaiveDateTime};
 use std::fs;
 use std::io;
 use tempfile::Builder;
@@ -29,43 +28,8 @@ async fn main() -> Result<()> {
 
     println!("{}", item_path.to_str().unwrap());
 
-    let mut rdr = csv::ReaderBuilder::new()
-        .delimiter(b';')
-        .from_path(item_path)?;
-
-    let mut measurement_vec = Vec::new();
-    for result in rdr.records() {
-        let record = result?;
-        let date_string = record.get(1).unwrap().to_string();
-        let year = date_string
-            .chars()
-            .take(4)
-            .collect::<String>()
-            .parse::<i32>()?;
-        let month = date_string
-            .chars()
-            .skip(4)
-            .take(2)
-            .collect::<String>()
-            .parse::<u32>()?;
-        let day = date_string
-            .chars()
-            .skip(6)
-            .take(2)
-            .collect::<String>()
-            .parse::<u32>()?;
-        let hour = date_string
-            .chars()
-            .skip(8)
-            .take(2)
-            .collect::<String>()
-            .parse::<u32>()?;
-
-        measurement_vec.push(data_access::TemperatureMeasurement::new(
-            NaiveDate::from_ymd(year, month, day).and_hms(hour, 0, 0),
-            record.get(3).unwrap().trim().parse::<f32>()?,
-        ));
-    }
+    let measurement_vec = data_access::load_data(item_path)?;
+    println!("{}", measurement_vec[0]);
 
     Ok(())
 }
